configurable buffer sizes
use the following command line options to configure buffers globally for all nodes for which there is no local configuration in the xml file:
 --interface-buffer=N   // network interface receive buffer in bytes
 --socket-recv-buffer=N // socket receive buffer initial size in bytes
 --socket-send-buffer=N // socket send buffer initial size in bytes

Note that the socket buffer values represent an initial size for TCP connections. TCP auto-tuning will further adjust the sizes based on both peers in the connection.

Similarly, each of these options can be configured on a per-node basis by specifying these configs as node attributes in the hosts.xml or similar file:
 <node interfacebuffer="1024000" socketrecvbuffer="256000" socketsendbuffer="128000" ... >

Note that these values will be optional for node elements.

refs #132